+++ webview-ui/src/components/line.tsx
@@ -14,573 +14,10 @@ import {
   createUnknown,
   prependToArray,
 } from "../lib/editionHelpers";
-import { AvailableDeclaration, findDeclarationsInScope } from "../lib/findDeclarations";
-
-// Valid C types for the type selector
-const C_TYPES = [
-  "void",
-  "int",
-  // "char",
-  // "float",
-  // "double",
-  // "short",
-  // "long",
-  // "unsigned char",
-  // "unsigned int",
-  // "unsigned short",
-  // "unsigned long",
-  // "signed char",
-  // "signed int",
-  // "signed short",
-  // "signed long",
-  // "long long",
-  // "unsigned long long",
-  // "size_t",
-  // "ptrdiff_t",
-];
-
-// ============================================================================
-// Generic Autocomplete Field Component
-// ============================================================================
-
-interface AutocompleteOption<T> {
-  value: T;
-  label: string;
-  key: string;
-  description?: React.ReactNode;
-  onSelect: (value: T) => void;
-}
-
-interface AutocompleteFieldProps<T> {
-  // Current value display
-  currentValue: string;
-  // In case current value is empty
-  placeholder: string;
-
-  // Available options
-  options: AutocompleteOption<T>[];
-
-  // Callbacks
-  onFocus?: () => void;
-  onNoMatch?: (inputText: string) => void; // Called when no valid option matches
-
-  // Focus management
-  focusRequest: { nodeId: string; fieldKey: string } | null;
-  nodeId: string;
-  fieldKey: string;
-  clearFocusRequest: () => void;
-
-  // Styling
-  className?: string;
-  isSelected?: boolean;
-
-  // Mode
-  readOnly?: boolean;
-}
-
-function AutocompleteField<T>({
-  currentValue,
-  placeholder,
-  options,
-  onFocus,
-  onNoMatch,
-  focusRequest,
-  nodeId,
-  fieldKey,
-  clearFocusRequest,
-  className,
-  isSelected = false,
-  readOnly = false,
-}: AutocompleteFieldProps<T>) {
-  const [inputValue, setInputValue] = React.useState(currentValue);
-  const [showDropdown, setShowDropdown] = React.useState(false);
-  const [selectedIndex, setSelectedIndex] = React.useState(-1);
-  const inputRef = React.useRef<HTMLInputElement>(null);
-  const hasFocusedRef = React.useRef(false);
-
-  // Filter options based on input
-  const filteredOptions = options.filter((option) =>
-    option.label.toLowerCase().includes(inputValue.toLowerCase())
-  );
-
-  // Find best match (exact prefix match, then contains match)
-  const getBestMatch = (): AutocompleteOption<T> | null => {
-    const exact = filteredOptions.find((option) =>
-      option.label.toLowerCase().startsWith(inputValue.toLowerCase())
-    );
-    return exact || filteredOptions[0] || null;
-  };
-
-  // Update input value when current value changes
-  React.useEffect(() => {
-    setInputValue(currentValue);
-  }, [currentValue]);
-
-  // Handle focus requests
-  React.useEffect(() => {
-    if (
-      focusRequest &&
-      focusRequest.nodeId === nodeId &&
-      focusRequest.fieldKey === fieldKey &&
-      !hasFocusedRef.current
-    ) {
-      console.log("Focusing autocomplete field for node:", nodeId, " field:", fieldKey);
-      if (inputRef.current) {
-        inputRef.current.focus();
-        inputRef.current.select();
-        hasFocusedRef.current = true;
-        clearFocusRequest();
-      }
-    }
-    if (!focusRequest) {
-      hasFocusedRef.current = false;
-    }
-  }, [focusRequest, nodeId, fieldKey, clearFocusRequest]);
-
-  const commitValue = (option: AutocompleteOption<T> | null) => {
-    if (option) {
-      // Call the option's onSelect callback
-      option.onSelect(option.value);
-      setInputValue(option.label);
-    } else {
-      // No match found - call onNoMatch if provided, otherwise revert
-      if (onNoMatch) {
-        onNoMatch(inputValue);
-      }
-      setInputValue(currentValue);
-    }
-    setShowDropdown(false);
-    setSelectedIndex(-1);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      if (showDropdown) {
-        e.stopPropagation();
-        if (selectedIndex >= 0) {
-          commitValue(filteredOptions[selectedIndex]);
-        } else {
-          const bestMatch = getBestMatch();
-          commitValue(bestMatch);
-        }
-      }
-    } else if (e.key === "Escape") {
-      if (showDropdown) {
-        e.stopPropagation();
-      }
-      setShowDropdown(false);
-      setSelectedIndex(-1);
-      setInputValue(currentValue);
-    } else if (e.key === "ArrowDown") {
-      e.preventDefault();
-      if (!showDropdown) {
-        setShowDropdown(true);
-        setSelectedIndex(0);
-      } else {
-        setSelectedIndex((prev) => Math.min(prev + 1, filteredOptions.length - 1));
-      }
-    } else if (e.key === "ArrowUp") {
-      e.preventDefault();
-      if (showDropdown) {
-        setSelectedIndex((prev) => Math.max(prev - 1, -1));
-      }
-    }
-    // if key is letter/number and dropdown not shown, show it
-    else if (e.key.length === 1 && !e.ctrlKey && !e.metaKey && !e.altKey) {
-      setShowDropdown(true);
-    }
-  };
-
-  const handleFocus = () => {
-    if (onFocus) {
-      onFocus();
-    }
-  };
-
-  const handleBlur = () => {
-    const bestMatch = getBestMatch();
-    commitValue(bestMatch);
-  };
-  const placeholderText = currentValue.length === 0 ? placeholder : currentValue;
-  const width = `${inputValue.length === 0 ? placeholderText.length : inputValue.length}ch`;
-
-  // Display description of first or selected option
-  const selectedOption: AutocompleteOption<T> | undefined =
-    (selectedIndex >= 0 ? filteredOptions[selectedIndex] : filteredOptions[0]) ?? undefined;
-
-  return (
-    <div style={{ position: "relative", display: "inline-block" }}>
-      <input
-        ref={inputRef}
-        className={`inline-editor ${className ?? ""}`}
-        style={{
-          ...(isSelected
-            ? {
-                backgroundColor: "rgba(255, 255, 255, 0.05)",
-                boxShadow: "inset 0 -1px 0 0 rgba(163, 209, 252, 0.5)",
-              }
-            : {}),
-          width,
-        }}
-        value={inputValue}
-        onChange={(e) => {
-          setInputValue(e.target.value);
-          setShowDropdown(true);
-          setSelectedIndex(-1);
-        }}
-        onKeyDown={handleKeyDown}
-        onFocus={handleFocus}
-        onBlur={handleBlur}
-        placeholder={placeholderText}
-        readOnly={readOnly}
-      />
-      {showDropdown && filteredOptions.length > 0 && !readOnly && (
-        <div
-          style={{
-            position: "absolute",
-            top: "100%",
-            left: "-2px",
-            zIndex: 1000,
-          }}
-        >
-          <div
-            style={{
-              position: "absolute",
-              top: "-3px",
-              display: "flex",
-              flexDirection: "row",
-              alignItems: "flex-start",
-            }}
-          >
-            <ScrollableBox>
-              {filteredOptions.map((option, index) =>
-                ((option: AutocompleteOption<T>, selected: boolean, index: number) => (
-                  <div
-                    key={option.key}
-                    style={{
-                      cursor: "pointer",
-                      backgroundColor: selected
-                        ? "var(--vscode-list-activeSelectionBackground)"
-                        : "transparent",
-                      color: selected ? "var(--vscode-list-activeSelectionForeground)" : "inherit",
-                      padding: "0px 2px",
-                    }}
-                    onMouseDown={(e) => {
-                      e.preventDefault();
-                      commitValue(option);
-                    }}
-                    onMouseEnter={() => setSelectedIndex(index)}
-                  >
-                    {option.label}
-                  </div>
-                ))(option, index === selectedIndex, index)
-              )}
-            </ScrollableBox>
-            {selectedOption.description && (
-              <ScrollableBox>{selectedOption.description}</ScrollableBox>
-            )}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ScrollableBox({ children }: { children: React.ReactNode }) {
-  return (
-    <div
-      style={{
-        background: "var(--vscode-dropdown-background)",
-        boxShadow: "inset 0 0 0 1px var(--vscode-dropdown-border)",
-        borderRadius: "2px",
-        padding: "1px",
-      }}
-    >
-      <div
-        style={{
-          maxHeight: "10rem",
-          minWidth: "10rem",
-          maxWidth: "20rem",
-          overflowY: "auto",
-          scrollbarGutter: "stable",
-        }}
-      >
-        {children}
-      </div>
-    </div>
-  );
-}
-
-// ============================================================================
-// Type Selector (uses AutocompleteField)
-// ============================================================================
-
-interface TypeSelectorProps<T extends objects.LanguageObject, K extends string & keyof T> {
-  node: T;
-  key: K;
-  parentInfo: ParentInfo;
-  className?: string;
-}
-
-function TypeSelector<T extends objects.LanguageObject, K extends string & keyof T>({
-  node,
-  key,
-  parentInfo,
-  className,
-}: TypeSelectorProps<T, K>) {
-  const {
-    selectedNodeId,
-    selectedKey,
-    onEdit,
-    setSelectedNodeId,
-    setSelectedKey,
-    setParentNodeInfo,
-    focusRequest,
-    clearFocusRequest,
-    mode,
-  } = useLineContext();
-
-  const isSelected = selectedNodeId === node.id && selectedKey && selectedKey === key;
-  const currentValue = String(node[key] ?? "");
-
-  function commitValue(selectedType: string) {
-    if (currentValue !== selectedType) {
-      node[key] = selectedType as T[K];
-      onEdit(node, key);
-    }
-  }
-
-  // Convert C_TYPES to AutocompleteOption format
-  const options: AutocompleteOption<string>[] = C_TYPES.map((type) => ({
-    value: type,
-    label: type,
-    key: type,
-    onSelect: (selectedType: string) => {
-      commitValue(selectedType);
-    },
-  }));
-
-  const handleNoMatch = (inputText: string) => {
-    // Check if input is a valid type despite not being in filtered options
-    const isValidType = C_TYPES.includes(inputText);
-    if (isValidType) {
-      commitValue(inputText);
-    } else if (inputText.trim() === "") {
-      // Empty input - revert to current value (do nothing)
-    } else {
-      // Invalid type - revert to current value (do nothing)
-    }
-  };
-
-  const handleFocus = () => {
-    setSelectedKey(key);
-    setSelectedNodeId(node.id);
-    setParentNodeInfo(parentInfo);
-  };
-
-  return (
-    <AutocompleteField
-      currentValue={currentValue}
-      placeholder="type"
-      options={options}
-      onNoMatch={handleNoMatch}
-      onFocus={handleFocus}
-      focusRequest={focusRequest}
-      nodeId={node.id}
-      fieldKey={key}
-      clearFocusRequest={clearFocusRequest}
-      className={className}
-      isSelected={!!isSelected}
-      readOnly={mode === "view"}
-    />
-  );
-}
-
-// ============================================================================
-// Reference Selector (uses AutocompleteField)
-// ============================================================================
-
-interface ReferenceSelectorProps {
-  node: objects.Reference;
-  parentInfo: ParentInfo;
-  callbacks?: NodeCallbacks;
-  className?: string;
-}
-
-function ReferenceSelector({ node, parentInfo, className, callbacks }: ReferenceSelectorProps) {
-  const {
-    onEdit,
-    setSelectedNodeId,
-    setSelectedKey,
-    setParentNodeInfo,
-    focusRequest,
-    clearFocusRequest,
-    mode,
-    nodeMap,
-    parentMap,
-  } = useLineContext();
-
-  const [options, setOptions] = React.useState<AutocompleteOption<AvailableDeclaration>[]>([]);
-
-  // Get the current identifier from the target declaration
-  const targetDecl = nodeMap.get(node.declarationId);
-  const currentIdentifier =
-    targetDecl && "identifier" in targetDecl ? String(targetDecl.identifier) : "";
-
-  const handleFocus = () => {
-    console.log("Focusing ReferenceSelector for node:", node.id);
-    setSelectedKey("declarationId");
-    setSelectedNodeId(node.id);
-    setParentNodeInfo(parentInfo);
-
-    // Find available declarations and convert to options
-    const declarations = findDeclarationsInScope(node, parentMap);
-    const newOptions: AutocompleteOption<AvailableDeclaration>[] = declarations.map((decl) => ({
-      value: decl,
-      label: decl.identifier,
-      description: (
-        <span style={{ fontStyle: "italic", color: "var(--vscode-descriptionForeground)" }}>
-          {decl.type}
-        </span>
-      ),
-      key: decl.id,
-      onSelect: (selectedDecl: AvailableDeclaration) => {
-        switch (selectedDecl.type) {
-          case "functionDeclaration":
-          case "functionDefinition": {
-            // Replace Reference with CallExpression
-            console.log(
-              "Converting Reference to CallExpression for function:",
-              selectedDecl.identifier
-            );
-
-            const newCallExpression: objects.CallExpression = {
-              id: crypto.randomUUID(),
-              type: "callExpression",
-              idDeclaration: selectedDecl.id,
-              identifier: selectedDecl.identifier,
-              argumentList: [],
-            };
-
-            nodeMap.set(newCallExpression.id, newCallExpression);
-            nodeMap.delete(node.id);
-
-            // Replace in parent
-            if (callbacks?.onReplace) {
-              callbacks.onReplace(node, newCallExpression);
-            }
-            break;
-          }
-          case "declaration":
-          case "functionParameter":
-            // Update reference to point to variable/parameter
-            node.declarationId = selectedDecl.id;
-            onEdit(node, "declarationId");
-        }
-      },
-    }));
-    setOptions(newOptions);
-    console.log("Available declarations for reference:", declarations);
-  };
-
-  const isSelected = focusRequest?.nodeId === node.id && focusRequest?.fieldKey === "declarationId";
-
-  return (
-    <AutocompleteField
-      currentValue={currentIdentifier}
-      placeholder="reference_name"
-      options={options}
-      onFocus={handleFocus}
-      focusRequest={focusRequest}
-      nodeId={node.id}
-      fieldKey="declarationId"
-      clearFocusRequest={clearFocusRequest}
-      className={className}
-      isSelected={!!isSelected}
-      readOnly={mode === "view"}
-    />
-  );
-}
-
-// ============================================================================
-// Call Expression Selector (uses AutocompleteField, only shows functions)
-// ============================================================================
-
-interface CallExpressionSelectorProps {
-  node: objects.CallExpression;
-  parentInfo: ParentInfo;
-  className?: string;
-}
-
-function CallExpressionSelector({ node, parentInfo, className }: CallExpressionSelectorProps) {
-  const {
-    onEdit,
-    setSelectedNodeId,
-    setSelectedKey,
-    setParentNodeInfo,
-    focusRequest,
-    clearFocusRequest,
-    mode,
-    nodeMap,
-    parentMap,
-  } = useLineContext();
-
-  const [options, setOptions] = React.useState<AutocompleteOption<AvailableDeclaration>[]>([]);
-
-  // Get the current identifier from the target declaration
-  const targetDecl = nodeMap.get(node.idDeclaration);
-  const currentIdentifier =
-    targetDecl && "identifier" in targetDecl ? String(targetDecl.identifier) : "";
-
-  const handleFocus = () => {
-    setSelectedKey("idDeclaration");
-    setSelectedNodeId(node.id);
-    setParentNodeInfo(parentInfo);
-
-    // Find available declarations and filter for functions only, then convert to options
-    const declarations = findDeclarationsInScope(node, parentMap);
-    const functions = declarations.filter(
-      (decl) => decl.type === "functionDeclaration" || decl.type === "functionDefinition"
-    );
-    const newOptions: AutocompleteOption<AvailableDeclaration>[] = functions.map((decl) => ({
-      value: decl,
-      label: decl.identifier,
-      description: (
-        <span style={{ fontStyle: "italic", color: "var(--vscode-descriptionForeground)" }}>
-          {decl.type}
-        </span>
-      ),
-      key: decl.id,
-      onSelect: (selectedDecl: AvailableDeclaration) => {
-        // Update the call expression to point to the new function
-        node.idDeclaration = selectedDecl.id;
-        node.identifier = selectedDecl.identifier;
-        onEdit(node, "idDeclaration");
-      },
-    }));
-    setOptions(newOptions);
-    console.log("Available functions for call expression:", functions);
-  };
-
-  const isSelected = focusRequest?.nodeId === node.id && focusRequest?.fieldKey === "idDeclaration";
-
-  return (
-    <AutocompleteField
-      currentValue={currentIdentifier}
-      placeholder="function_name"
-      options={options}
-      onFocus={handleFocus}
-      focusRequest={focusRequest}
-      nodeId={node.id}
-      fieldKey="idDeclaration"
-      clearFocusRequest={clearFocusRequest}
-      className={className}
-      isSelected={!!isSelected}
-      readOnly={mode === "view"}
-    />
-  );
-}
+import { TypeSelector } from "./selectors/TypeSelector";
+import { ReferenceSelector } from "./selectors/ReferenceSelector";
+import { CallExpressionSelector } from "./selectors/CallExpressionSelector";
+import { EditableField } from "./EditableField";
 
 // Hook to handle focus requests for structural nodes (nodes with tabIndex={0})
 function useFocusStructuralNode(nodeId: string) {
@@ -609,101 +46,6 @@ function useFocusStructuralNode(nodeId: string) {
   }, [focusRequest, nodeId, clearFocusRequest]);
 
   return nodeRef;
-}
-
-interface EditableFieldProps<T extends objects.LanguageObject, K extends string & keyof T> {
-  node: T;
-  key: K;
-  parentInfo: ParentInfo;
-  className?: string;
-  placeholder: string;
-}
-
-function EditableField<T extends objects.LanguageObject, K extends string & keyof T>({
-  node,
-  key,
-  parentInfo,
-  className,
-  placeholder,
-}: EditableFieldProps<T, K>) {
-  const {
-    selectedNodeId,
-    selectedKey,
-    onEdit,
-    setSelectedNodeId,
-    setSelectedKey,
-    setParentNodeInfo,
-    focusRequest,
-    clearFocusRequest,
-    mode,
-  } = useLineContext();
-  const isSelected = selectedNodeId === node.id && selectedKey && selectedKey === key;
-  const initialValue = String(node[key] ?? "");
-  const [inputValue, setInputValue] = React.useState(initialValue);
-  const inputRef = React.useRef<HTMLInputElement>(null);
-  const hasFocusedRef = React.useRef(false);
-
-  React.useEffect(() => {
-    setInputValue(String(node[key] ?? ""));
-  }, [node, key]);
-
-  // React to focus requests - only focus once per request
-  React.useEffect(() => {
-    if (
-      focusRequest &&
-      focusRequest.nodeId === node.id &&
-      focusRequest.fieldKey === key &&
-      !hasFocusedRef.current
-    ) {
-      console.log("Focusing input for node:", node.id, " key:", key);
-      if (inputRef.current) {
-        inputRef.current.focus();
-        inputRef.current.select();
-        hasFocusedRef.current = true;
-        // Clear the focus request after handling it
-        clearFocusRequest();
-      }
-    }
-    // Reset the flag when focus request changes
-    if (!focusRequest) {
-      hasFocusedRef.current = false;
-    }
-  }, [focusRequest, node.id, key, clearFocusRequest]);
-
-  // width in ch units, at least 1ch
-  const width =
-    (inputValue.length === 0 ? placeholder.length : Math.max(1, inputValue.length)) + "ch";
-
-  return (
-    <input
-      ref={inputRef}
-      className={`inline-editor ${className ?? ""}`}
-      style={{
-        ...(isSelected
-          ? {
-              backgroundColor: "rgba(255, 255, 255, 0.05)",
-              boxShadow: "inset 0 -1px 0 0 rgba(163, 209, 252, 0.5)",
-            }
-          : {}),
-        width, // dynamically set width in ch
-      }}
-      value={inputValue}
-      placeholder={placeholder}
-      onChange={(e) => setInputValue(e.target.value)}
-      onFocus={() => {
-        setSelectedKey(key);
-        setSelectedNodeId(node.id);
-        setParentNodeInfo(parentInfo);
-      }}
-      onBlur={() => {
-        if (initialValue !== inputValue) {
-          node[key] = inputValue as T[K];
-          onEdit(node, key);
-        }
-      }}
-      readOnly={mode === "view"}
-    />
-  );
 }
 
 interface ObjectProps {
